# src/main/resources/api-specs/consumer-api-specs.yml
openapi: 3.0.3
info:
  title: Space Cats Marketplace API
  version: 1.0.0
  description: API for managing products in the marketplace
servers:
  - url: http://localhost:8080
paths:
  /api/products:
    get:
      summary: Get all products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
    post:
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDto'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          description: Validation error
  /api/products/{id}:
    get:
      summary: Get a product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '404':
          description: Product not found
    put:
      summary: Update a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequestDto'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        '400':
          description: Validation error
        '404':
          description: Product not found
    delete:
      summary: Delete a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Product deleted
        '404':
          description: Product not found
components:
  schemas:
    ProductDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        inStock:
          type: integer
    ProductRequestDto:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 50
          description: Must contain a cosmic word
        description:
          type: string
          maxLength: 200
        price:
          type: number
          format: double
          minimum: 0
        inStock:
          type: integer
          minimum: 0